WELD-1787 Apply MethodInjectionPoint optimization to all callables
