+++ impl/src/main/java/org/jboss/weld/injection/AbstractCallableInjectionPoint.java
@@ -24,11 +24,13 @@ import java.lang.reflect.Type;
 import java.util.List;
 import java.util.Set;
 
+import javax.enterprise.inject.TransientReference;
 import javax.enterprise.inject.spi.AnnotatedCallable;
 import javax.enterprise.inject.spi.Bean;
 import javax.enterprise.inject.spi.InjectionPoint;
 
 import org.jboss.weld.annotated.enhanced.EnhancedAnnotatedCallable;
+import org.jboss.weld.annotated.enhanced.EnhancedAnnotatedParameter;
 import org.jboss.weld.exceptions.UnsupportedOperationException;
 import org.jboss.weld.injection.attributes.SpecialParameterInjectionPoint;
 import org.jboss.weld.injection.attributes.WeldInjectionPointAttributes;
@@ -42,6 +44,7 @@ abstract class AbstractCallableInjectionPoint<T, X, S extends Member> implements
     private final Bean<?> declaringBean;
     private final List<ParameterInjectionPoint<?, X>> parameters;
     private final Set<InjectionPoint> injectionPoints;
+    protected final boolean hasTransientReferenceParameter;
 
     protected AbstractCallableInjectionPoint(EnhancedAnnotatedCallable<T, X, S> callable, Bean<?> declaringBean, Class<?> declaringComponentClass, boolean observerOrDisposer, InjectionPointFactory factory, BeanManagerImpl manager) {
         this.declaringBean = declaringBean;
@@ -56,6 +59,16 @@ abstract class AbstractCallableInjectionPoint<T, X, S extends Member> implements
                 }
             };
         }
+        this.hasTransientReferenceParameter = initHasTransientReference(callable.getEnhancedParameters());
+    }
+
+    private static boolean initHasTransientReference(List<? extends EnhancedAnnotatedParameter<?, ?>> parameters) {
+        for (EnhancedAnnotatedParameter<?, ?> parameter : parameters) {
+            if (parameter.isAnnotationPresent(TransientReference.class)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     @Override
